Use our fork of interpol to disable csrf middleware on the stub app
Again, this isn't a thing we would ever need to do, but it's nice for a
reader of the blog post to be able to click a link and see an endpoint
response from the stub app (rather than a 403).

Interpol doesn't expose any sinatra configuration, so we had to fork the
gem itself.